requirements.txt
acme==0.31.0
aioconsole==0.1.16
aiofiles==0.4.0
aiohttp==3.6.2
aioopenssl==0.5.1
aiosasl==0.4.0
aioxmpp==0.11.0
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.0
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.8.24
chardet==3.0.4
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.6.1
discord==1.0.1
discord.py==1.3.3
dnspython==1.16.0
entrypoints==0.3
fortnitepy==2.3
future==0.16.0
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.5
httplib2==0.11.3
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.6
josepy==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
lxml==4.5.0
mock==2.0.0
multidict==4.7.5
mysql-connector==2.2.9
parsedatetime==2.4
pbr==4.2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pycurl==7.43.0.2
PyGObject==3.30.4
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-apt==1.8.4.1
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2019.1
pyxdg==0.25
reportbug===7.5.3-deb10u1
requests==2.21.0
requests-toolbelt==0.8.0
rfc3986==1.4.0
sanic==19.12.2
SecretStorage==2.3.1
six==1.12.0
sniffio==1.1.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==2.0
tzlocal==1.5.1
ujson==2.0.3
urllib3==1.24.1
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
psutil~=5.7.0